feat(pragma_exex): Cleaning
