+++ crates/client/rpc/src/versions/v0_7_1/methods/read/estimate_fee.rs
@@ -5,11 +5,11 @@ use starknet_core::types::{BlockId, BroadcastedTransaction, FeeEstimate, Simulat
 use mc_exec::ExecutionContext;
 use mp_transactions::broadcasted_to_blockifier;
 
-use crate::utils::ResultExt;
 use crate::versions::v0_7_1::methods::trace::trace_transaction::FALLBACK_TO_SEQUENCER_WHEN_VERSION_BELOW;
 use crate::Starknet;
 use mp_rpc::errors::StarknetRpcApiError;
 use mp_rpc::errors::StarknetRpcResult;
+use mp_rpc::utils::ResultExt;
 
 /// Estimate the fee associated with transaction
 ///
