+++ crates/client/rpc/src/versions/v0_7_1/methods/read/get_transaction_receipt.rs
@@ -3,8 +3,8 @@ use starknet_core::types::{Felt, TransactionFinalityStatus, TransactionReceiptWi
 
 use mp_rpc::errors::{StarknetRpcApiError, StarknetRpcResult};
 
-use crate::utils::ResultExt;
 use crate::Starknet;
+use mp_rpc::utils::ResultExt;
 
 /// Get the transaction receipt by the transaction hash.
 ///
